First draft of customization command
In this first draft, we do not have a full implementation of JSONPath
for use in a Customizer's Replacements.  For now the syntax and
semantics of a Replacement's Path are what is implemented in
https://github.com/tidwall/sjson - with one nod toward JSONPath.  The
Replacement's Path must start with "$.", and this is stripped off and
the remainder of the path is handled by sjson.

Signed-off-by: Mike Spreitzer <mspreitz@us.ibm.com>