Make executed docs not fetch repo
It is enough that the caller does that.  Letting the executed doc work
with the context in which it finds itself makes the executed doc a
useful test, for users and for CI.

Also: make the cleanup of leftover junk in the pre-reqs.md not kill
the `make` running it.

Also: slow down at the end of example1, because I got a failure
without it.

Signed-off-by: Mike Spreitzer <mspreitz@us.ibm.com>