coordinate between multiple parties which share one kcp's admin.kubeconfig
The issue that's handled here is due to the fact the kube-bind only has a CLI but lacks API for programmatic access. Because of this, we need scripts which calls `kubectl-bind` for each of the workspaces.
For IMWs and WMWs, we can control at what time to call the scripts in the generated 'example one' main test script. But for MBWs, we don't fully controll at what time the scritps are called, because those
calls are issued by the mailbox-controller. Futhermore, the mailbox controller has concurrent workers, and has retries.

The test script and the mailbox controller workers shares a single admin.kubeconfig. And they all WRITE to the same admin.kubeconfig by changing the workspaces. (In contrast, in the containerized
deployment, each container has a separate admin.kubeconfig). Therefore, the workspace is uncertain for the command lines in the test! No supprise that the test is real flaky.

This is hard to debug. Until I see complaints about a 'admin.kubeconfig.lock' file, I didn't realize there are concurrent access to admin.kubeconfig.
I added more time after mailbox controller starts to allow its workers to call the scripts and finish the bindings.