Delete controller_test.go because it is pointless
This test was never fully drafted because I realized that it has a
problem.

The fakes do not implement ServerPreferredResources (i.e., implement
it by returning nothing), which is the form of discovery that drives
the binding controller. To make this a useful test we would need to
either create and use better fakes or change the controller to use the
other form of discovery and take the necessary pains to fake it
through the existing fakes.

Signed-off-by: Mike Spreitzer <mspreitz@us.ibm.com>